requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
faster-whisper==1.0.3
pyannote.audio==3.1.1
numpy>=1.24
soundfile==0.12.1
